app/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pyannote.audio
huggingface-hub
requests
torch
torchaudio
soundfile
